Teach the script to handle drive letters
When passing the class path to Java, we separate the individual class
path elements with colons, as is the convention with Unix. That means
that the class path elements themselves cannot contain colons.
Therefore, just replace `C:` with `/C` so that MSys/MSys2 can handle the
class path properly before passing it to `java.exe`.

With Cygwin, the conversion should be `C:` -> `/cygdrive/C`, but then,
Cygwin's `dirname` does not return the path using a drive letter to
begin with.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>